kafka-producer/requirements.txt
yfinance
kafka-python
pandas
